move EventsOrMetricsRequest to app/components/charts
